requirements.txt
# requirements.txt
tqdm
PyYAML